Added the option to "pre-start" a line
When starting to draw a new line, often the first movement is absorbed by slack in the
mechanism and in the motors themselves, especially when the direction of movement
reverses the previous one.

The new pre_start=True option for drawing and plotting methods positions the pen .5cm
before the point at which it will start drawing, so that by the time the pen is
lowered the mechanism will have already moved a small amount in the right direction.

The aim is to take up any slack.