Implemented a more nuanced sophisticated_angles_to_pulse_widths
Previously, sophisticated_angles_to_pulse_widths took a mean hysteresis value
across the pulse-width ranges, and applied that after the numpy-based
angles-to-pulse-widths calculation was complete.

Now, two numpy functions are obtained for each servo, one for
ascending angles and one for descending. When a pulse-width value is
required from a sophisticated_angles_to_pulse_widths method, a direction
argument ("asc" or "des") can be provided. If not provided, the mean
result of both ascending and descending functions will be returned.